DIGITAL DC MULTIMETER CONSTRUCTION
Project Objective:
The objective of this project is to create a simple digital multimeter prototype on a printed circuit
board using basic electronic components. The aim is to measure voltage, current, and resistance
values through the use of two probes and a switch in the circuit.
Project Scenario:
The switch will be set to the required position for the data to be measured. The ends of the two
probes will be touched to the connection points of the current, resistance, or voltage to be measured.
As a result, the measured data will be displayed on the LCD screen.